# dashboard/.schema-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Orchestrator Dashboard",
        "description": "REST/SSE API of the YAOE-FLOW observability dashboard.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/agents/{id}/harness": {
            "get": {
                "operationId": "getApiAgentsByIdHarness",
                "tags": [
                    "Agents"
                ],
                "summary": "List the agent's harness configs",
                "responses": {
                    "200": {
                        "description": "Configs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true
                    }
                ]
            }
        },
        "/api/agents/{id}/harness/{harnessId}": {
            "put": {
                "operationId": "putApiAgentsByIdHarnessByHarnessId",
                "tags": [
                    "Agents"
                ],
                "summary": "Update a harness config",
                "responses": {
                    "200": {
                        "description": "Config atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        },
                                        "config": {
                                            "type": "object",
                                            "propertyNames": {
                                                "type": "string"
                                            },
                                            "additionalProperties": {}
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "config"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "model": {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "null"
                                            }
                                        ]
                                    },
                                    "settingsJson": {
                                        "type": "string"
                                    },
                                    "mcpServersJson": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "harnessId",
                        "required": true
                    }
                ]
            },
            "delete": {
                "operationId": "deleteApiAgentsByIdHarnessByHarnessId",
                "tags": [
                    "Agents"
                ],
                "summary": "Delete a harness config",
                "responses": {
                    "200": {
                        "description": "Removida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        }
                                    },
                                    "required": [
                                        "ok"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "harnessId",
                        "required": true
                    }
                ]
            }
        },
...
        "/api/harness/{id}/models": {
            "get": {
                "operationId": "getApiHarnessByIdModels",
                "tags": [
                    "Harness"
                ],
                "summary": "Models accepted by the harness (from the detection cache)",
                "responses": {
                    "200": {
                        "description": "Modelos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "harnessId": {
                                            "type": "string"
                                        },
                                        "modelSelection": {
                                            "type": "string"
                                        },
                                        "models": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "propertyNames": {
                                                    "type": "string"
                                                },
                                                "additionalProperties": {}
                                            }
                                        },
                                        "defaultModelId": {
                                            "anyOf": [
                                                {
                                                    "type": "string"
                                                },
                                                {
                                                    "type": "null"
                                                }
                                            ]
                                        },
                                        "checkedAt": {
                                            "anyOf": [
                                                {
                                                    "type": "string"
                                                },
                                                {
                                                    "type": "null"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "harnessId",
                                        "modelSelection",
                                        "models"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Harness desconhecido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ]
            }
        },
        "/api/harness/{id}/redetect": {
            "post": {
                "operationId": "postApiHarnessByIdRedetect",
                "tags": [
                    "Harness"
                ],
                "summary": "Re-detect a specific harness",
                "responses": {
                    "200": {
                        "description": "Detec\u00e7\u00e3o atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Harness desconhecido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ]
            }
        },
        "/api/harness/{id}/budgets": {
            "put": {
                "operationId": "putApiHarnessByIdBudgets",
                "tags": [
                    "Harness"
                ],
                "summary": "Update harness budgets",
                "responses": {
                    "200": {
                        "description": "Budgets atualizados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        },
                                        "budgets": {
                                            "type": "object",
                                            "propertyNames": {
                                                "type": "string"
                                            },
                                            "additionalProperties": {}
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "budgets"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Harness desconhecido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "dailyLimit": {
                                        "type": "number"
                                    },
                                    "weeklyLimit": {
                                        "type": "number"
                                    },
                                    "monthlyLimit": {
                                        "type": "number"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "enum": [
                                            "usd",
                                            "tokens"
                                        ]
                                    },
                                    "action": {
                                        "type": "string",
                                        "enum": [
                                            "avisar",
                                            "pausar"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ]
            }
        },
...
    },
...
}